+++ backend/server.js
@@ -12,21 +12,26 @@ const app = express();
 app.use(cors());
 app.use(express.json({ limit: '50mb' }));
 
-connectToDB();
-seedData();
-
 // Log all requests
 app.use((req, res, next) => {
     console.log(`${req.method} ${req.url}`);
     next();
 });
 
-app.use('/api/wardrobe', wardrobeRoutes);
+const startServer = async () => {
+    await connectToDB();
+    await seedData();
 
-// Test route
-app.get('/api/test', (req, res) => {
-    res.json({ message: 'API is working!' });
-});
+    app.use('/api/wardrobe', wardrobeRoutes);
+    console.log("Routes are supported")
+
+    // Test route
+    app.get('/api/test', (req, res) => {
+        res.json({ message: 'API is working!' });
+    });
+}
 
 const PORT = process.env.PORT || 3000;
 app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
+
+startServer();
